requirements.txt
python-dotenv 
python-telegram-bot 
aiohttp
gspread 
gspread-asyncio 
google-auth-oauthlib
